Add Rust code launcher test
This C stub launches preemptible Rust jobs.  Unfortunately, it turns out that it leaks anything for
which the Rust runtime defers cleanup via the at_exit mechanism; even a call to Rust's exit() (which
would eventually need to be seccomp()'d or linked such that it didn't really kill the process)
doesn't cause the cleanup to occur.
The easiest fix will probably to make the whole thing a Rust program so that all setup/teardown
happen when the program starts and ends.