launcher.rs: Memoize the loaded LibFun instances to avoid linear dlopen() searches
It turns out that dlopen()'s RTLD_NODELETE option is crippled by the fact that the implementation
takes O(n) time to determine whether the library with a given path has already been loaded! Instead,
we now memoize the two-pointer LibFun entries in a hash table; this offers the added theoretical
advantage of later being able to unload libraries.