Makefile: Linkage fixup: Automatically detect bogus shared libraries
Bogus shared libraries are those that the compiler will *never* choose to link against because they
don't provide any unresolved symbols at link-time.  This patch obviates the need to write per-target
warning rules for libraries for which declaring a dynamic dependency has confusing, build-breaking
results if a stale static version of the library already exists in the build directory.